PolyBoom: Use size_t

+++ src/graphics/effects/PolyBoom.cpp
@@ -283,11 +283,11 @@ void PolyBoomAddSplat(const Sphere & sp, const Color3f & col, long flags) {
 			if((ep->type & POLY_TRANS) && !(ep->type & POLY_WATER))
 				continue;
 			
-			long nbvert = (ep->type & POLY_QUAD) ? 4 : 3;
+			size_t nbvert = (ep->type & POLY_QUAD) ? 4 : 3;
 			
 			bool oki = false;
 			
-			for(long k = 0; k < nbvert; k++) {
+			for(size_t k = 0; k < nbvert; k++) {
 				
 				if(PointIn2DPolyXZ(&TheoricalSplat, ep->v[k].p.x, ep->v[k].p.z)
 				   && glm::abs(ep->v[k].p.y - py) < 100.f) {
@@ -334,7 +334,7 @@ void PolyBoomAddSplat(const Sphere & sp, const Color3f & col, long flags) {
 					pb.timecreation = now;
 					pb.rgb = col;
 					
-					for(int k = 0; k < nbvert; k++) {
+					for(size_t k = 0; k < nbvert; k++) {
 						
 						float vdiff = glm::abs(ep->v[k].p.y - RealSplatStart.y);
